services/dashboard/requirements.txt
kafka-python
psycopg2-binary
python-dotenv
pandas
streamlit
